Move upload into helper function
